Initial Commit
The UI is a mess currently because I haven't completely wrapped my head
around constraints and Auto Layout. The delete is fully functional. Each
picButton (i.e. scissors, toilet) has the same name as the object that
it is and has the picButtonPressed action.
So, to sum up, we need:
- paging system
- better UI design
- ability to adapt to different sized screens
- icons
- long press support for subcategory control